Refactored the sample project
