+++ sample/build.gradle
@@ -3,9 +3,8 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 22
     buildToolsVersion "22.0.1"
-
     defaultConfig {
-        applicationId "no.jlagos.viewpagerindicatortest"
+        applicationId 'no.jlagos.viewpagerindicator'
         minSdkVersion 14
         targetSdkVersion 22
         versionCode 1
@@ -17,10 +16,11 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    productFlavors {
+    }
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
     compile 'com.android.support:appcompat-v7:22.1.1'
     compile 'com.jakewharton.timber:timber:3.1.0'
     compile project(':viewpageindicator')
