+++ sample/src/main/java/no/jlagos/sample/MainActivity.java
@@ -1,4 +1,4 @@
-package no.jlagos.viewpagerindicatortest;
+package no.jlagos.sample;
 
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
