+++ sample/src/main/res/.gitignore
