+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':viewpageindicator'
+include ':sample', ':viewpageindicator'
